move database to own file
